backend/requirements.txt
# Core Django
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Multi-tenant
django-tenants==3.6.0
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-otp==1.3.0
django-allauth==0.60.1

# Database & Caching
redis==5.0.1
celery==5.3.4
django-cacheops==7.2

# API Documentation
drf-spectacular==0.27.0

# AI & ML
langchain==0.1.0
langchain-openai==0.0.5
pgvector==0.2.4
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File Handling
Pillow==10.1.0
django-storages==1.14.2

# Email & Notifications
django-anymail==10.2

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
django-sslserver==0.22

# Production
gunicorn==21.2.0
whitenoise==6.6.0
